Extract the TableRenderer rendering seam contract
Introduce TableRenderer<TData> and TableRendererProps<TData> — a small,
UI-library-agnostic contract describing what a table renderer is handed
(one page of rows, selection, paging-agnostic props). Move the row-click
event into its own file so the contract can reference it, and re-express
DataTableCoreProps as `extends TableRendererProps` plus the PrimeReact
pass-through and rendering extras. DataTableCore is unchanged at runtime;
it is now the default implementation of the seam.
